requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
alembic==1.18.4
anyio==4.13.0
apache-airflow==2.9.1
apache-airflow-providers-common-compat==1.6.1
apache-airflow-providers-common-io==1.5.4
apache-airflow-providers-common-sql==1.26.0
apache-airflow-providers-fab==1.5.3
apache-airflow-providers-ftp==3.12.3
apache-airflow-providers-http==5.2.2
apache-airflow-providers-imap==3.8.3
apache-airflow-providers-smtp==2.0.3
apache-airflow-providers-sqlite==4.0.2
apispec==6.10.0
argcomplete==3.6.3
asgiref==3.11.1
async-timeout==5.0.1
attrs==26.1.0
babel==2.18.0
blinker==1.9.0
cachelib==0.13.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.2
connexion==2.14.2
cron_descriptor==2.0.8
croniter==6.2.2
cryptography==46.0.6
Deprecated==1.3.1
dill==0.4.1
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.1
Flask==2.2.5
Flask-AppBuilder==4.5.3
Flask-Babel==2.0.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.12
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
frozenlist==1.8.0
fsspec==2026.3.0
google-re2==1.1.20251105
googleapis-common-protos==1.74.0
greenlet==3.0.3
grpcio==1.80.0
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_metadata==8.7.1
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lazy-object-proxy==1.12.0
limits==5.8.0
linkify-it-py==2.1.0
lockfile==0.12.2
Mako==1.3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
marshmallow-oneofschema==3.2.0
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.5.0
mdurl==0.1.2
methodtools==0.4.7
more-itertools==11.0.1
multidict==6.7.1
numpy==2.2.6
opentelemetry-api==1.40.0
opentelemetry-exporter-otlp==1.40.0
opentelemetry-exporter-otlp-proto-common==1.40.0
opentelemetry-exporter-otlp-proto-grpc==1.40.0
opentelemetry-exporter-otlp-proto-http==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
ordered-set==4.1.0
packaging==26.0
pandas==2.2.2
pathlib_abc==0.5.2
pathspec==1.0.4
pendulum==3.2.0
playwright==1.43.0
pluggy==1.6.0
prison==0.2.1
propcache==0.4.1
protobuf==6.33.6
psutil==7.2.2
pycparser==3.0
pyee==11.1.0
Pygments==2.20.0
PyJWT==2.12.1
PyMySQL==1.1.0
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.33.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.2
rpds-py==0.30.0
setproctitle==1.3.7
six==1.17.0
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.42.1
sqlparse==0.5.5
tabulate==0.10.0
tenacity==9.1.4
termcolor==3.3.0
text-unidecode==1.3
tomli==2.4.1
typing_extensions==4.15.0
tzdata==2026.1
uc-micro-py==2.0.0
unicodecsv==0.14.1
universal_pathlib==0.3.10
urllib3==2.6.3
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==2.1.2
WTForms==3.2.1
yarl==1.23.0
zipp==3.23.0